protocol: fix logging of errors with coroutines
future.exception() does not set sys.exc_info, so we cannot build a
JsonRpcInternalError from it and we cannot use logger.exception.

This problem was observable with a coroutine like this:

  @server.feature(TEXT_DOCUMENT_DID_OPEN)
  async def did_open(
      ls: LanguageServer,
      params: DidOpenTextDocumentParams,
  ) -> None:
    raise ValueError()

the logged information was:

  pygls.protocol ERROR    Exception in notification: {'code': -32602, 'message': 'None: None', 'data': "{'traceback': []}"}
  NoneType: None

Use the exception returned from future.exception() to get a proper
traceback in the log.