types: fix WorkDoneProgressOptions.workDoneProgress type
It seems there was a confusing with WorkDoneProgressParams and
WorkDoneProgressOptions.

[WorkDoneProgressParams](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workDoneProgressParams)
is the type used for the token in methods and notifications, it is the
token itself, so it has type str | int. This type is correct, this was
to explain context.

[WorkDoneProgressOptions](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workDoneProgressOptions)
is the type used for registration, it is used to indicate if work done
progress are supported, so it is a boolean. This type was wrong.

It seems as of today vscode does include workDoneProgress in requests:
 - https://github.com/microsoft/vscode-languageserver-node/issues/528#issuecomment-541600565
 - https://github.com/microsoft/vscode/issues/105870