progress: support cancellation from client
This allows servers to react when client cancel requests, either by
polling is_cancelled or by registering a callback using on_cancel.

The old cancel method is removed, because it was not corresponding to
anything from the spec.

Note that the tests have to be skipped on pyodide, as they need
asyncio.sleep for a delay non blocking message processing.